+++ apps/client/src/app/order/form/page.tsx
@@ -1,70 +1,13 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
+import ReactQueryProvider from '@/providers/QueryProvider';
 
-import BottomButton from '@/components/common/BottomButton';
-import AddressSection from '@/components/features/order/form/AddressSection';
-import AgreementSection from '@/components/features/order/form/AgreementSection';
-import PriceSummarySection from '@/components/features/order/form/PriceSummarySection';
-import OrderItemsSection from '@/components/features/order/OrderItemsSection';
-import Header from '@/components/layouts/Header';
-import { useSelectedAddressStore } from '@/store/useSelectedAddressStore';
+import OrderFormContent from './OrderFormPageContent';
 
-export default function Page() {
-    const { selectedAddress } = useSelectedAddressStore();
-
-    const [deliveryRequest, setDeliveryRequest] = useState<string | null>(null);
-    const [textareaContent, setTextareaContent] = useState('');
-
-    const [hasAgreedToRequiredTerms, setHasAgreedToRequiredTerms] =
-        useState(false);
-
-    const isDisabled = !(selectedAddress && hasAgreedToRequiredTerms);
-    const router = useRouter();
-
+export default function OrderFormPage() {
     return (
-        <div className="pb-17">
-            <Header type="title" title="주문/결제" />
-
-            <div className="flex flex-col px-5 pt-2">
-                {/* 배송지 정보 */}
-                <div className="flex flex-col gap-3 border-t border-black pb-8 pt-4">
-                    <AddressSection
-                        deliveryRequest={deliveryRequest}
-                        setDeliveryRequest={setDeliveryRequest}
-                        textareaContent={textareaContent}
-                        setTextareaContent={setTextareaContent}
-                        addressInfo={selectedAddress}
-                    />
-                </div>
-
-                {/* 주문 상품 */}
-                <div className="flex flex-col gap-3 border-t border-black pb-8 pt-4">
-                    <OrderItemsSection variant="default" />
-                </div>
-
-                {/* 결제 금액 */}
-                <div className="flex flex-col border-t border-black pb-5 pt-5">
-                    <PriceSummarySection />
-                </div>
-
-                {/* 약관 동의 */}
-                <div className="flex flex-col border-t border-black pb-5 pt-5">
-                    <AgreementSection
-                        setHasAgreedToRequiredTerms={
-                            setHasAgreedToRequiredTerms
-                        }
-                    />
-                </div>
-            </div>
-
-            {/* 버튼 */}
-            <BottomButton
-                text={'21,000원 결제하기'}
-                isDisabled={isDisabled}
-                onClick={() => router.push('/order/confirm')}
-            />
-        </div>
+        <ReactQueryProvider>
+            <OrderFormContent />
+        </ReactQueryProvider>
     );
 }
