+++ apps/client/src/providers/QueryProvider.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
